requirements.txt
streamlit
flirimageextractor
matplotlib
numpy
opencv-python
Pillow
flyr
scipy
paramiko # posso remover depois
seaborn